requirements.txt
gradio>=3.0
python-docx>=0.8.11
langchain>=0.0.300
openai>=1.0.0
faiss-cpu>=1.7.4
chromadb>=0.3.28
tiktoken
sentence-transformers
numpy
